refactor(character): move grant api to adminController
