+++ src/main/java/online/lifeasgame/character/application/result/PlayerResult.java
@@ -52,36 +52,4 @@ public class PlayerResult {
             );
         }
     }
-
-    public record CurrentHp(
-            int value
-    ) {
-        public static CurrentHp from(Player player) {
-            return new CurrentHp(player.getHealth().current());
-        }
-    }
-
-    public record HpCapacity(
-            int cap
-    ) {
-        public static HpCapacity from(Player player) {
-            return new HpCapacity(player.getHealth().cap());
-        }
-    }
-
-    public record CurrentMp(
-            int value
-    ) {
-        public static CurrentMp from(Player player) {
-            return new CurrentMp(player.getMana().current());
-        }
-    }
-
-    public record MpCapacity(
-            int cap
-    ) {
-        public static MpCapacity from(Player player) {
-            return new MpCapacity(player.getMana().cap());
-        }
-    }
 }
