+++ src/main/java/online/lifeasgame/character/presentation/response/PlayerResponse.java
@@ -64,36 +64,4 @@ public class PlayerResponse {
             }
         }
     }
-
-    public record CurrentHp(
-            int currentHp
-    ) {
-        public static CurrentHp of(int currentHp) {
-            return new CurrentHp(currentHp);
-        }
-    }
-
-    public record HpCapacity(
-            int hpCapacity
-    ) {
-        public static HpCapacity of(int hpCapacity) {
-            return new HpCapacity(hpCapacity);
-        }
-    }
-
-    public record CurrentMp(
-            int currentMp
-    ) {
-        public static CurrentMp of(int currentMp) {
-            return new CurrentMp(currentMp);
-        }
-    }
-
-    public record MpCapacity(
-            int mpCapacity
-    ) {
-        public static MpCapacity of(int mpCapacity) {
-            return new MpCapacity(mpCapacity);
-        }
-    }
 }
